인터페이스 기능 초안 정의

+++ Study/src/StudentManagementSystem/io/SortedStudent.java
@@ -1,6 +1,6 @@
 package StudentManagementSystem.io;
 
 public interface SortedStudent {
-    void sort(double average); //성적 순 정렬
+    void sort(float average); //성적 순 정렬
     void sort(String sno); // 학번 순 정렬
 }
